feat(protocol): 协议核心 Mod 分端二进制桥 + 无头测试基线
- com.game.protocol：IProtocolHost 传输抽象，FNV-1a 确定性 msgId
- MirrorNetModule：Mirror 传输与协议 Mod 桥接
- tests/TestRunner：契约/ECS/协议/推箱子无头测试（17 项）
- setup-mirror.sh / verify-unity.sh：环境脚本

+++ mods/com.game.protocol/src/ProtocolClient.cs
@@ -23,7 +23,7 @@ namespace Com.Game.Protocol
         }
 
         /// <summary>客户端 Mod 注册协议解析器（编解码器）。</summary>
-        public ushort Register<T>(ModId owner, IMessageCodec<T> codec) where T : struct
+        public uint Register<T>(ModId owner, IMessageCodec<T> codec) where T : struct
             => _core.Register(owner, codec);
 
         /// <summary>上行：客户端 Mod 消息 → 序列化 → 二进制 → 服务端。</summary>
@@ -39,7 +39,7 @@ namespace Com.Game.Protocol
             => _core.Handle(owner, handler);
 
         /// <summary>查询消息类型对应的 ID（未注册返回 0）。</summary>
-        public ushort IdOf<T>() where T : notnull
+        public uint IdOf<T>() where T : notnull
             => _core.IdOf<T>();
     }
 }
